Use type hieararchy from AST for GlympseKit classes' hierarchy and protocols.
The generated wrapper now looks to the protocol list for GlyEventSink to
determine whether to apply the sink template. This can be made more
generic when we handle boilerplate for other protocols.

Unhandled protocols can be dealt with by exluding the protocl generation
and using an empty protocol definition. The protocol marker still has
some value.